Seperate annotations to NavigationMapView extension file.
